Better table format in dry run
