+++ config/random-generator/deployment/random-generator-2.yml
@@ -6,9 +6,10 @@ metadata:
     scaler/state-peak-replicas: "5"
     scaler/state-bau-replicas: "2"
     scaler/state-default-replicas: "1"
+    scaler/rapid-scaling: "false"
   labels:
     app: random-generator-2
-    scaler/opt-in: "false"
+    scaler/opt-in: "true"
 spec:
   progressDeadlineSeconds: 20
   selector:
