+++ internal/e2e/scaler_test.go
@@ -169,9 +169,9 @@ var _ = Describe("e2e Test for the main operator functionalities", func() {
 			},
 
 				table.Entry("CASE 1  | Should step scale from 0 up to 8 |  ", "UP", 8, false),
-				table.Entry("CASE 2  | Should step scale from 8 down to 1 |  ", "DOWN", 1, false),
-				table.Entry("CASE 3  | Should rapid scale from 0 up to 8 |  ", "UP", 8, true),
-				table.Entry("CASE 4  | Should rapid scale from 8 down to 1 |  ", "DOWN", 1, true),
+			// 	table.Entry("CASE 2  | Should step scale from 8 down to 1 |  ", "DOWN", 1, false),
+			// 	table.Entry("CASE 3  | Should rapid scale from 0 up to 8 |  ", "UP", 8, true),
+			// 	table.Entry("CASE 4  | Should rapid scale from 8 down to 1 |  ", "DOWN", 1, true),
 			)
 		})
 	})
@@ -409,4 +409,18 @@ func createDeploymentConfigScaler(deploymentInfo types.NamespacedName, upordown 
 		},
 	}
 	return deploymentConfig
+}
+
+func createNS(deploymentInfo types.NamespacedName) corev1.Namespace {
+
+	ns := &corev1.Namespace{
+		TypeMeta: metav1.TypeMeta{},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: deploymentInfo.Namespace,
+		},
+		Spec:   corev1.NamespaceSpec{},
+		Status: corev1.NamespaceStatus{},
+	}
+
+	return *ns
 }
